+++ example/cumulus-tf/main.tf
@@ -137,7 +137,7 @@ module "cumulus" {
   sync_granule_s3_jitter_max_ms     = var.sync_granule_s3_jitter_max_ms
 
   # For update-granules-cmr-metadata-file-links-task in order to update granule identifiers or not
-  update_granule_identifiers = var.default_update_granule_identifiers
+  update_granule_identifiers = var.update_granule_identifiers
 
   launchpad_api         = var.launchpad_api
   launchpad_certificate = var.launchpad_certificate